Move expandable-panels dir
