+++ components/expandable-panels/PrimaryExpandablePanel.module.scss
